Fix all golangci-lint issues across codebase
- Fix predeclared: rename variables shadowing built-ins (max, min, new, print)
- Fix unparam: mark unused parameters with _ or add nolint for API consistency
- Fix nilerr: add nolint comments for intentional error skipping in WalkDir
- Fix nilnil: add nolint comments for valid nil returns in YAML parsing
- Fix unconvert: remove unnecessary type conversions in free_unix.go, ls_unix.go
- Fix mirror: use bytes.IndexByte instead of strings.IndexByte in xargs.go
- Fix prealloc: preallocate slices with known capacity in ls.go, awk.go, jq.go
- Fix recvcheck: add nolint for bubbletea interface requiring value receivers
- Fix gocritic: convert if-else chains to switch statements where appropriate
- Fix maintidx: add nolint for test function with expected high complexity
- Disable gocritic.ifElseChain globally in .golangci.yml
- Simplify boolean expression in uuid.go for hex character validation

+++ pkg/cli/free_unix.go
@@ -23,6 +23,7 @@ func getMemInfo() (MemInfo, error) {
 		scanner := bufio.NewScanner(file)
 		for scanner.Scan() {
 			line := scanner.Text()
+
 			fields := strings.Fields(line)
 			if len(fields) < 2 {
 				continue
@@ -59,13 +60,13 @@ func getMemInfo() (MemInfo, error) {
 	}
 
 	unit := uint64(sysinfo.Unit)
-	info.MemTotal = uint64(sysinfo.Totalram) * unit
-	info.MemFree = uint64(sysinfo.Freeram) * unit
+	info.MemTotal = sysinfo.Totalram * unit
+	info.MemFree = sysinfo.Freeram * unit
 	info.MemAvailable = info.MemFree // Approximation
-	info.Buffers = uint64(sysinfo.Bufferram) * unit
+	info.Buffers = sysinfo.Bufferram * unit
 	info.Cached = 0 // Not available from sysinfo
-	info.SwapTotal = uint64(sysinfo.Totalswap) * unit
-	info.SwapFree = uint64(sysinfo.Freeswap) * unit
+	info.SwapTotal = sysinfo.Totalswap * unit
+	info.SwapFree = sysinfo.Freeswap * unit
 
 	return info, nil
 }
